pyproject.toml
[project]
name = "airborne"
version = "0.1.0"
description = "Blind-accessible flight simulator with self-voicing and realistic physics"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "AirBorne Team"}
]
dependencies = [
    "pygame>=2.5.0",
    "PyBASS3>=0.2.0",
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
    "pyfmodex>=0.7.2",
    "pyttsx3>=2.99",
    "pydub>=0.25.1",
    "mypy>=1.18.2",
    "pylint>=4.0.0",
    "ruff>=0.14.0",
    "websockets>=15.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pylint>=3.0.0",
    "types-pyyaml>=6.0.0",
    "pre-commit>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/airborne"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_unimported = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
check_untyped_defs = true
strict_equality = true

[tool.pylint.messages_control]
max-line-length = 100
disable = [
    "C0111",  # missing-docstring (we'll enforce this manually)
    "R0903",  # too-few-public-methods (dataclasses are fine)
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --strict-markers"

[dependency-groups]
dev = [
    "pre-commit>=4.3.0",
    "pyinstaller>=6.16.0",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]
